egl-wayland2: Fix libdrm loading when wl_drm is not available
If the compositor does not support wl_drm and we fail to get the
drmGetDeviceFromDevId symbol then the driver load will fail. This
happens with steam due to it hijacking dlopen/dlsym. This change
has us directly dlopen libdrm and get a symbol from it instead of
relying on RTLD_DEFAULT.